# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import bcrypt from "bcryptjs";
import jwt from "jsonwebtoken";
import { db } from "@workspace/db";
import { employeesTable, employeePinsTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { AuthWithPinBody } from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/employees/login-list", async (_req, res): Promise<void> => {
  const employees = await db
    .select({
      id: employeesTable.id,
      name: employeesTable.name,
      role: employeesTable.role,
    })
    .from(employeesTable)
    .where(eq(employeesTable.active, true))
    .orderBy(employeesTable.name);

  res.json(employees);
});

router.post("/auth/pin", async (req, res): Promise<void> => {
  const parsed = PinLoginInput.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Empleado y PIN son obligatorios" });
    return;
  }

  const { employeeId, pin } = parsed.data;

  const rows = await db
    .select({
      id: employeesTable.id,
      name: employeesTable.name,
      role: employeesTable.role,
      pinHash: employeePinsTable.pinHash,
    })
    .from(employeesTable)
    .innerJoin(employeePinsTable, eq(employeePinsTable.employeeId, employeesTable.id))
    .where(and(eq(employeesTable.id, employeeId), eq(employeesTable.active, true)));

  const employee = rows[0];
  if (!employee || !(await bcrypt.compare(pin, employee.pinHash))) {
    res.status(401).json({ error: "PIN incorrecto" });
    return;
  }

  const secret = process.env["SESSION_SECRET"];
  if (!secret) {
    res.status(500).json({ error: "Configuración de servidor incompleta" });
    return;
  }

  const token = jwt.sign(
    { id: employee.id, name: employee.name, role: employee.role },
    secret,
    { expiresIn: "12h" }
  );

  res.json({
    token,
    employee: { id: employee.id, name: employee.name, role: employee.role },
  });
});

export default router;