# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Piccolo TPV API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: string;
  name: string;
  role: string;
}

export interface PinLoginInput {
  employeeId: string;
  pin: string;
}

export interface AuthResponse {
  token: string;
  employee: Employee;
}

export interface Zone {
  id: string;
  name: string;
  type: string;
}

export interface Table {
  id: string;
  zoneId: string;
  name: string;
  capacity: number;
  status: string;
  x: number;
  y: number;
  shape: string;
}

export interface OrderItemModifier {
  id: string;
  orderItemId: string;
  modifierId?: string;
  modifierName: string;
  priceDelta: string;
}

export interface OrderItem {
  id: string;
  orderId: string;
  productId: string;
  productName: string;
  quantity: number;
  unitPrice: string;
  status: string;
  notes: string;
  allergyNote: string;
  hasAllergy: boolean;
  modifiers?: OrderItemModifier[];
  createdAt: string;
}

export interface Order {
  id: string;
  tableId: string;
  employeeId: string;
  status: string;
  createdAt: string;
  sentAt?: string;
  items?: OrderItem[];
}

export interface TableWithOrder {
  table: Table;
  order: Order;
}

export interface ModifierOption {
  id: string;
  name: string;
  priceDelta: string;
}

export interface ModifierGroup {
  id: string;
  name: string;
  required: boolean;
  maxSelect: number;
  modifiers: ModifierOption[];
}

export type UpdateOrderItemDetailsInputModifiersItem = {
  modifierId?: string;
  modifierName: string;
  priceDelta: string;
};

export interface UpdateOrderItemDetailsInput {
  notes?: string;
  allergyNote?: string;
  hasAllergy?: boolean;
  modifiers?: UpdateOrderItemDetailsInputModifiersItem[];
}

export interface WaiterNotification {
  id: string;
  employeeId: string;
  orderId?: string;
  type: string;
  title: string;
  message: string;
  readAt?: string;
  createdAt: string;
}

export interface Category {
  id: string;
  name: string;
  sortOrder: number;
}

export interface Product {
  id: string;
  categoryId: string;
  name: string;
  price: string;
  prepZone: string;
}

export interface KitchenTask {
  id: string;
  orderId: string;
  orderItemId: string;
  prepZone: string;
  productName: string;
  quantity: number;
  status: string;
  allergyNote: string;
  hasAllergy: boolean;
  tableName: string;
  employeeName: string;
  employeeId?: string;
  createdAt: string;
  updatedAt: string;
  readyAt?: string;
  collectedAt?: string;
  servedAt?: string;
}

export interface AddOrderItemInput {
  productId: string;
  quantity: number;
  notes?: string;
}

export type UpdateTaskStatusInputStatus = typeof UpdateTaskStatusInputStatus[keyof typeof UpdateTaskStatusInputStatus];


export const UpdateTaskStatusInputStatus = {
  new: 'new',
  preparing: 'preparing',
  ready: 'ready',
  collected: 'collected',
  served: 'served',
  cancelled: 'cancelled',
} as const;

export interface UpdateTaskStatusInput {
  status: UpdateTaskStatusInputStatus;
}

export type PaseActionInputAction = typeof PaseActionInputAction[keyof typeof PaseActionInputAction];


export const PaseActionInputAction = {
  collected: 'collected',
  served: 'served',
} as const;

export interface PaseActionInput {
  action: PaseActionInputAction;
}

export interface DashboardSummary {
  totalTables: number;
  occupiedTables: number;
  freeTables: number;
  totalZones: number;
}

export interface PaymentMethod {
  id: string;
  code: string;
  name: string;
  active: boolean;
  sortOrder: number;
}

export interface CashSession {
  id: string;
  employeeId: string;
  openedAt: string;
  openingFloat: string;
  closedAt?: string;
  expectedCash?: string;
  countedCash?: string;
  difference?: string;
  status: string;
}

export type CashSessionWithEmployee = CashSession & {
  employeeName?: string;
};

export type CashMovementMovementType = typeof CashMovementMovementType[keyof typeof CashMovementMovementType];


export const CashMovementMovementType = {
  in: 'in',
  out: 'out',
} as const;

export interface CashMovement {
  id: string;
  cashSessionId: string;
  movementType: CashMovementMovementType;
  amount: string;
  reason: string;
  employeeId: string;
  createdAt: string;
}

export type CashSessionSummarySalesByMethodItem = {
  methodCode: string;
  methodName: string;
  total?: string;
};

export interface CashSessionSummary {
  session: CashSession;
  salesByMethod: CashSessionSummarySalesByMethodItem[];
  movements: CashMovement[];
}

export interface OpenCashSessionInput {
  openingFloat?: string;
}

export type AddCashMovementInputMovementType = typeof AddCashMovementInputMovementType[keyof typeof AddCashMovementInputMovementType];


export const AddCashMovementInputMovementType = {
  in: 'in',
  out: 'out',
} as const;

export interface AddCashMovementInput {
  movementType: AddCashMovementInputMovementType;
  amount: string;
  reason: string;
}

export interface CloseCashSessionInput {
  countedCash: string;
}

export type PaymentSummaryOrder = {
  id: string;
  status: string;
  tableName?: string;
  employeeName?: string;
};

export type PaymentSummaryItemsItem = {
  id: string;
  productName: string;
  quantity: number;
  unitPrice: string;
  lineTotal: string;
};

export type PaymentSummaryPaymentsItem = {
  id: string;
  amount: string;
  status: string;
  methodCode: string;
  methodName: string;
  createdAt: string;
};

export interface PaymentSummary {
  order: PaymentSummaryOrder;
  items: PaymentSummaryItemsItem[];
  subtotal: string;
  taxTotal: string;
  total: string;
  paid: string;
  remaining: string;
  methods: PaymentMethod[];
  payments: PaymentSummaryPaymentsItem[];
}

export type AddPaymentInputMethodCode = typeof AddPaymentInputMethodCode[keyof typeof AddPaymentInputMethodCode];


export const AddPaymentInputMethodCode = {
  cash: 'cash',
  card: 'card',
  bizum: 'bizum',
} as const;

export interface AddPaymentInput {
  methodCode: AddPaymentInputMethodCode;
  amount: string;
  reference?: string;
}

export type PaymentResultPayment = {
  id: string;
  amount: string;
  status: string;
};

export type PaymentResultTicket = {
  id?: string;
  ticketNumber?: number;
  total?: string;
  issuedAt?: string;
} | null;

export interface PaymentResult {
  payment: PaymentResultPayment;
  ticket?: PaymentResultTicket;
  change: string;
  newRemaining: number;
}

export type TicketDataTicket = {
  id: string;
  orderId: string;
  ticketNumber: number;
  subtotal: string;
  taxTotal: string;
  total: string;
  issuedAt: string;
};

export type TicketDataOrder = {
  id: string;
  tableName?: string;
  createdAt?: string;
};

export type TicketDataItemsItem = {
  productName: string;
  quantity: number;
  unitPrice: string;
  lineTotal: string;
};

export type TicketDataPaymentsItem = {
  amount: string;
  methodName: string;
};

export interface TicketData {
  ticket: TicketDataTicket;
  order: TicketDataOrder;
  items: TicketDataItemsItem[];
  payments: TicketDataPaymentsItem[];
  employeeName?: string;
}

export interface ErrorResponse {
  error: string;
}