# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Piccolo TPV API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/pin:
    post:
      operationId: authWithPin
      tags: [auth]
      summary: Authenticate employee with PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinLoginInput"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /kds/{zone}:
    get:
      operationId: getKdsTasks
      tags: [kds]
      summary: Get kitchen tasks for a production zone
      security:
        - bearerAuth: []
      parameters:
        - name: zone
          in: path
          required: true
          schema:
            type: string
            enum: [cocina, pizza, ensalada, barra, pase]
      responses:
        "200":
          description: Kitchen tasks for the zone
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KitchenTask"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kitchen-tasks/{taskId}/status:
    patch:
      operationId: updateKitchenTaskStatus
      tags: [kds]
      summary: Update status of a kitchen task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStatusInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenTask"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cash-sessions/open:
    post:
      operationId: openCashSession
      tags: [cash]
      summary: Open a new cash session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenCashSessionInput"
      responses:
        "201":
          description: Cash session opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashSession"
        "409":
          description: Session already open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cash-sessions/current:
    get:
      operationId: getCurrentCashSession
      tags: [cash]
      summary: Get the currently open cash session (or null)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current cash session or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashSessionWithEmployee"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cash-sessions/{id}/movements:
    post:
      operationId: addCashMovement
      tags: [cash]
      summary: Add a manual cash movement (in or out)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCashMovementInput"
      responses:
        "201":
          description: Movement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashMovement"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found or closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cash-sessions/{id}/close:
    post:
      operationId: closeCashSession
      tags: [cash]
      summary: Close the cash session with counted cash
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseCashSessionInput"
      responses:
        "200":
          description: Session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashSession"
        "404":
          description: Session not found or already closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cash-sessions/{id}/summary:
    get:
      operationId: getCashSessionSummary
      tags: [cash]
      summary: Get detailed session summary with sales by method and movements
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashSessionSummary"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
      required: [id, name, role]
    PinLoginInput:
      type: object
      properties:
        employeeId:
          type: string
        pin:
          type: string
      required: [employeeId, pin]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
      required: [token, employee]
...
    KitchenTask:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        orderItemId:
          type: string
        prepZone:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        status:
          type: string
        allergyNote:
          type: string
        hasAllergy:
          type: boolean
        tableName:
          type: string
        employeeName:
          type: string
        employeeId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        readyAt:
          type: string
        collectedAt:
          type: string
        servedAt:
          type: string
      required: [id, orderId, orderItemId, prepZone, productName, quantity, status, allergyNote, hasAllergy, tableName, employeeName, createdAt, updatedAt]
...
    UpdateTaskStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [new, preparing, ready, collected, served, cancelled]
      required: [status]
...
    CashSession:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        openedAt:
          type: string
        openingFloat:
          type: string
        closedAt:
          type: string
        expectedCash:
          type: string
        countedCash:
          type: string
        difference:
          type: string
        status:
          type: string
      required: [id, employeeId, openedAt, openingFloat, status]
    CashSessionWithEmployee:
      allOf:
        - $ref: "#/components/schemas/CashSession"
        - type: object
          properties:
            employeeName:
              type: string
    CashMovement:
      type: object
      properties:
        id:
          type: string
        cashSessionId:
          type: string
        movementType:
          type: string
          enum: [in, out]
        amount:
          type: string
        reason:
          type: string
        employeeId:
          type: string
        createdAt:
          type: string
      required: [id, cashSessionId, movementType, amount, reason, employeeId, createdAt]
    CashSessionSummary:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/CashSession"
        salesByMethod:
          type: array
          items:
            type: object
            properties:
              methodCode:
                type: string
              methodName:
                type: string
              total:
                type: string
            required: [methodCode, methodName]
        movements:
          type: array
          items:
            $ref: "#/components/schemas/CashMovement"
      required: [session, salesByMethod, movements]
    OpenCashSessionInput:
      type: object
      properties:
        openingFloat:
          type: string
    AddCashMovementInput:
      type: object
      properties:
        movementType:
          type: string
          enum: [in, out]
        amount:
          type: string
        reason:
          type: string
      required: [movementType, amount, reason]
    CloseCashSessionInput:
      type: object
      properties:
        countedCash:
          type: string
      required: [countedCash]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]